build.sh --extra-args 로 config.js 만 업데이트할 수 있도록 함

+++ build.sh
@@ -10,6 +10,11 @@ set -e
 export $(egrep -v '^#' .env | xargs)
 eval "echo \"$(cat $CONFIG_TEMPLATE)\"" > $CONFIG
 
+if [ $# -ne 0 ]; then
+    echo "Extra arguments provided"
+    exit 1
+fi
+
 # build
 rm -fr ./.vuepress/dist && npm run build
